batch: Rename text file storage
Batch text persistence is the last writer left in the former storage module
after binary, gitlink, and entry operations moved to named modules.

Keeping that code under batch.storage hides that the module handles text
ownership publication. New callers cannot tell which storage module owns
text batch writes from the path name alone.

This commit completes the storage split by renaming storage.py to
text_file_storage.py and updating production imports.

Subsequent commits will move tests and architecture checks to the text
storage owner.
