+++ src/git_stage_batch/batch/text_file_storage.py
@@ -1,4 +1,4 @@
-"""Batch storage operations: file management."""
+"""Text file batch persistence."""
 
 from __future__ import annotations
