+++ src/git_stage_batch/commands/selection/batch_line_updates.py
@@ -8,7 +8,7 @@ from contextlib import ExitStack
 from ...batch.operations import create_batch
 from ...batch.query import read_batch_metadata
 from ...batch.source_refresh import acquire_batch_ownership_update_for_selection
-from ...batch.storage import add_file_to_batch
+from ...batch.text_file_storage import add_file_to_batch
 from ...batch.validation import batch_exists
 from ...data.file_modes import detect_file_mode
 from ...data.session import snapshot_file_if_untracked
