+++ src/git_stage_batch/commands/selection/selected_change_batch_discarding.py
@@ -10,7 +10,7 @@ from ...batch.display import annotate_with_batch_source
 from ...batch.operations import create_batch
 from ...batch.query import read_batch_metadata
 from ...batch.source_refresh import acquire_batch_ownership_update_for_selection
-from ...batch.storage import add_file_to_batch
+from ...batch.text_file_storage import add_file_to_batch
 from ...batch.validation import batch_exists
 from ...core.buffer import LineBuffer
 from ...core.diff_parser import (
